requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SSLify==0.1.5
get==2019.4.13
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.1
numpy==1.17.3
Pillow==6.2.1
post==2019.4.13
public==2019.4.13
PyYAML==5.1.2
query-string==2019.4.13
requests==2.22.0
scipy==1.3.1
six==1.13.0
urllib3==1.25.6
Werkzeug==0.16.0
